# specs/001-personal-growth-planner/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Growth Planner API
  description: |
    RESTful API for the Personal Growth Planner application.
    Supports OKR management, learning path tracking, calendar scheduling, and progress monitoring.
  version: 1.0.0
  contact:
    name: Agiemme Planner Team
    email: support@agiemme-planner.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.agiemme-planner.com/v1
    description: Production server
...
paths:
...
  /okrs/objectives:
    get:
      tags: [OKRs]
      summary: Get all objectives for authenticated user
      operationId: getObjectives
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by life area category
          schema:
            $ref: '#/components/schemas/ObjectiveCategory'
        - name: status
          in: query
          description: Filter by status
          schema:
            $ref: '#/components/schemas/ObjectiveStatus'
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, targetDate, updatedAt]
            default: targetDate
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of objectives with key results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ObjectiveWithKeyResults'
                  meta:
                    $ref: '#/components/schemas/ListMetadata'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [OKRs]
      summary: Create a new objective
      operationId: createObjective
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectiveRequest'
      responses:
        '201':
          description: Objective created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveWithKeyResults'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /okrs/objectives/{objectiveId}:
    get:
      tags: [OKRs]
      summary: Get objective by ID with all key results
      operationId: getObjectiveById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ObjectiveId'
      responses:
        '200':
          description: Objective details with key results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveWithKeyResults'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [OKRs]
      summary: Update objective
      operationId: updateObjective
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ObjectiveId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateObjectiveRequest'
      responses:
        '200':
          description: Objective updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveWithKeyResults'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [OKRs]
      summary: Delete objective (cascades to key results)
      operationId: deleteObjective
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ObjectiveId'
      responses:
        '204':
          description: Objective deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  parameters:
    ObjectiveId:
      name: objectiveId
      in: path
      required: true
      description: Unique objective identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ObjectiveCategory:
      type: string
      enum:
        - career
        - health
        - relationships
        - skills
        - financial
        - personal
    ObjectiveStatus:
      type: string
      enum:
        - active
        - completed
        - archived
        - abandoned
    KeyResultStatus:
      type: string
      enum:
        - not_started
        - in_progress
        - completed
        - at_risk
...
    Objective:
      type: object
      required:
        - id
        - userId
        - title
        - category
        - targetDate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 5
          maxLength: 200
          example: Advance my career in software engineering
        description:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/ObjectiveCategory'
        targetDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ObjectiveStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ObjectiveWithKeyResults:
      allOf:
        - $ref: '#/components/schemas/Objective'
        - type: object
          required:
            - keyResults
          properties:
            keyResults:
              type: array
              minItems: 2
              maxItems: 5
              items:
                $ref: '#/components/schemas/KeyResult'
    KeyResult:
      type: object
      required:
        - id
        - objectiveId
        - description
        - targetValue
        - currentValue
        - unit
        - deadline
        - completionPercentage
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        objectiveId:
          type: string
          format: uuid
        description:
          type: string
          minLength: 10
          maxLength: 300
          example: Complete 3 certifications
        targetValue:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          example: 3.00
        currentValue:
          type: number
          format: decimal
          minimum: 0
          example: 1.00
        unit:
          type: string
          minLength: 1
          maxLength: 50
          example: certifications
        deadline:
          type: string
          format: date
        completionPercentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
          description: Calculated (currentValue / targetValue) * 100
        status:
          $ref: '#/components/schemas/KeyResultStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateObjectiveRequest:
      type: object
      required:
        - title
        - category
        - targetDate
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        description:
          type: string
        category:
          $ref: '#/components/schemas/ObjectiveCategory'
        targetDate:
          type: string
          format: date
        keyResults:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: object
            required:
              - description
              - targetValue
              - unit
              - deadline
            properties:
              description:
                type: string
                minLength: 10
                maxLength: 300
              targetValue:
                type: number
                format: decimal
                minimum: 0
                exclusiveMinimum: true
              unit:
                type: string
                minLength: 1
                maxLength: 50
              deadline:
                type: string
                format: date
    UpdateObjectiveRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
        description:
          type: string
        category:
          $ref: '#/components/schemas/ObjectiveCategory'
        targetDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ObjectiveStatus'
...
    ListMetadata:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        count:
          type: integer
          description: Number of items in current response
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - validationErrors
          properties:
            message:
              type: string
              example: Validation failed
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name that failed validation
                  message:
                    type: string
                    description: Validation error message
  responses:
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'